EventReader: fix value returned by "count" request
It was returning the number of the "last" event to read. This could be
different than the number of events in the file when the service is
configured with the "skip" and/or "max" options.

Make it always return the total number of events in the file, and use a
separate variable to track the "last" event that must be read.